+++ railties/test/generators/migration_generator_test.rb
@@ -183,6 +183,80 @@ class MigrationGeneratorTest < Rails::Generators::TestCase
     end
   end
 
+  def test_should_create_migration_to_add_index
+    migration = "add_index_author_id_on_books"
+    run_generator [migration, "author_id"]
+
+    assert_migration "db/migrate/#{migration}.rb" do |content|
+      assert_match(/AddIndexAuthorIdOnBooks/, content, "wrong class name")
+      assert_method :change, content do |change|
+        assert_match(/add_index :books, :author_id/, change, "index not present")
+      end
+    end
+  end
+
+  def test_should_create_migration_to_remove_index
+    migration = "remove_index_author_id_from_books"
+    run_generator [migration, "author_id"]
+
+    assert_migration "db/migrate/#{migration}.rb" do |content|
+      assert_match(/RemoveIndexAuthorIdFromBooks/, content, "wrong class name")
+      assert_method :change, content do |change|
+        assert_match(/remove_index :books, column: :author_id/, change, "index not present")
+      end
+    end
+  end
+
+  def test_should_create_migration_to_add_index_uniq
+    migration = "add_index_author_id_uniq_on_books"
+    run_generator [migration, "author_id:uniq"]
+
+    assert_migration "db/migrate/#{migration}.rb" do |content|
+      assert_match(/AddIndexAuthorIdUniqOnBooks/, content, "wrong class name")
+      assert_method :change, content do |change|
+        assert_match(/add_index :books, :author_id, unique: true/, change, "index not present")
+      end
+    end
+  end
+
+  def test_should_create_migration_to_remove_index_uniq
+    migration = "remove_index_author_id_uniq_from_books"
+    run_generator [migration, "author_id:uniq"]
+
+    assert_migration "db/migrate/#{migration}.rb" do |content|
+      assert_match(/RemoveIndexAuthorIdUniqFromBooks/, content, "wrong class name")
+      assert_method :change, content do |change|
+        assert_match(/remove_index :books, column: :author_id, unique: true/, change, "index not present")
+      end
+    end
+  end
+
+  def test_should_create_migration_to_add_multiple_indexes
+    migration = "add_indexes_author_id_uniq_and_library_id_on_books"
+    run_generator [migration, "author_id:uniq", "library_id"]
+
+    assert_migration "db/migrate/#{migration}.rb" do |content|
+      assert_match(/AddIndexesAuthorIdUniqAndLibraryIdOnBooks/, content, "wrong class name")
+      assert_method :change, content do |change|
+        assert_match(/add_index :books, :author_id, unique: true/, change, "index author_id not present")
+        assert_match(/add_index :books, :library_id/, change, "index library_id not present")
+      end
+    end
+  end
+
+  def test_should_create_migration_to_remove_multiple_indexes_uniq
+    migration = "remove_indexes_author_id_uniq_and_library_id_from_books"
+    run_generator [migration, "author_id:uniq", "library_id"]
+
+    assert_migration "db/migrate/#{migration}.rb" do |content|
+      assert_match(/RemoveIndexesAuthorIdUniqAndLibraryIdFromBooks/, content, "wrong class name")
+      assert_method :change, content do |change|
+        assert_match(/remove_index :books, column: :author_id, unique: true/, change, "index author_id not present")
+        assert_match(/remove_index :books, column: :library_id/, change, "index library_id not present")
+      end
+    end
+  end
+
   def test_should_create_empty_migrations_if_name_not_start_with_add_or_remove_or_create
     migration = "delete_books"
     run_generator [migration, "title:string", "content:text"]